fix libcanard compilation & CAN types

+++ src_pico/src/task_can.h
@@ -3,5 +3,5 @@
 #include "Arduino.h"
 
 void taskCAN(void *pvParameters);
-void onReceiveCanPacket(int packetSize, uint32_t packetId, uint8_t *packetData,
+void onReceiveCanPacket(uint8_t packetSize, uint32_t packetId, uint8_t *packetData,
                         bool extended);
